[bugfix] 修复编译错误

+++ sdk/ecmedia/win/video_capturer/video_capturer.cc
@@ -25,8 +25,8 @@ VideoCapturer* VideoCapturer::CreateCaptureDevice(
   std::unique_ptr<webrtc::VideoCaptureModule::DeviceInfo> device_info(
       webrtc::VideoCaptureFactory::CreateDeviceInfo());
 
-  index = index >= device_info->NumberOfDevices()
-              ? device_info->NumberOfDevices()-1
+  index = index >= (int)device_info->NumberOfDevices()
+              ? (int)device_info->NumberOfDevices()-1
                  : index;
   char device_name[256];
   char unique_name[256];
